requirements.txt
ytmusicapi
yt-dlp
pyyaml
